Add `SearchController`
